% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muestra_contribuyentes.R
\name{muestra_contribuyentes}
\alias{muestra_contribuyentes}
\title{Muestra de los Países de las y los Contribuyentes}
\usage{
muestra_contribuyentes(n = 500)
}
\arguments{
\item{n}{Un número entero no negativo que indica el número de elementos a
muestrear.}
}
\description{
Toma una muestra, con reemplazo, de los países de las y los contribuyentes
de este repositorio.
}
\examples{
# Tomamos una muestra de 10 banderas de las y los contribuyentes.
muestra_contribuyentes(10)

}
