% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_contribuyentes.R
\name{obtener_contribuyentes}
\alias{obtener_contribuyentes}
\title{Obtener Contribuyentes}
\usage{
obtener_contribuyentes()
}
\description{
Devuelve un tibble que lista todos los contribuyentes de este repositorio.
}
\examples{
head(obtener_contribuyentes())
}
